+++ tests/Feature/NumberTest.php
@@ -3,7 +3,10 @@ namespace MarcoConsiglio\BCMathExtended\Tests\Feature;
 
 use MarcoConsiglio\BCMathExtended\Number;
 use BcMath\Number as BcMathNumber;
+use MarcoConsiglio\BCMathExtended\Range;
 use MarcoConsiglio\BCMathExtended\Tests\TestCaseWithDataProviders;
+use MarcoConsiglio\FakerPhpNumberHelpers\NextFloat;
+use Override;
 use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\Attributes\TestDox;
 use RoundingMode;
@@ -457,6 +460,137 @@ class NumberTest extends TestCaseWithDataProviders
         $this->assertTrue($number_A->gte($number_B), "$number_A ≱ $number_B");
     }
 
+    #[TestDox("can check if it is in range.")]
+    public function test_inRange(): void
+    {
+        /**
+         * In Range
+         */
+        // Arrange
+        $range = new Range(-100, 100);
+        $number = $this->randomNumber(
+            min: $range->start->toFloat(),
+            max: $range->end->toFloat()
+        );
+
+        // Act & Assert
+        $this->assertTrue(
+            $number->inRange($range),
+            "$number is not inside $range->start/$range->end range."    
+        );
+
+        /**
+         * Out of range
+         */
+        // Arrange
+        $number = self::$faker->randomElement([
+            $this->randomNumber(max: NextFloat::before($range->start->toFloat())),
+            $this->randomNumber(min: NextFloat::after($range->end->toFloat()))
+        ]);
+
+        // Act & Assert
+        $this->assertFalse(
+            $number->inRange($range),
+            "$number is inside $range->start/$range->end range."    
+        );
+    }
+
+    #[TestDox("can check if it is in range excluded its start.")]
+    public function test_inRangeMinExcluded(): void
+    {
+        /**
+         * In range
+         */
+        // Arrange
+        $range = new Range(-100, 100);
+        $number = $this->randomNumber(
+            min: NextFloat::after($range->start->toFloat()),
+            max: $range->end->toFloat()
+        );
+
+        // Act & Assert
+        $this->assertTrue(
+            $number->inRangeMinExcluded($range),
+            "$number is not inside $range->start/$range->end range."    
+        ); 
+
+        /**
+         * Out of range
+         */
+        // Arrange
+        $number = $this->randomNumber(max: $range->start->toFloat());
+
+        // Act & Assert
+        $this->assertFalse(
+            $number->inRangeMinExcluded(new Range($range->start, $range->end)),
+            "$number is inside $range->start/$range->end range."
+        );
+    }
+
+    #[TestDox("can check if it is in range excluded its end.")]
+    public function test_inRangeMaxExcluded(): void
+    {
+        /**
+         * In range
+         */
+        // Arrange
+        $range = new Range(-100, 100);
+        $number = $this->randomNumber(
+            min: $range->start->toFloat(),
+            max: $range->end->sub(new Number(1))->toFloat()
+        );
+
+        // Act & Assert
+        $this->assertTrue(
+            $number->inRangeMaxExcluded(new Range($range->start, $range->end)),
+            "$number is not inside $range->start/$range->end range."    
+        );
+
+        /**
+         * Out of range
+         */
+        // Arrange
+        $number = $this->randomNumber(max: NextFloat::before($range->start->toFloat()));
+
+        // Act & Assert
+        $this->assertFalse(
+            $number->inRangeMaxExcluded($range),
+            "$number is inside $range->start/$range->end range."
+        );
+    }
+
+    #[TestDox("can check if it is in range excluded both its extremes.")]
+    public function test_inRangeExtremesExcluded(): void
+    {
+        /**
+         * In range
+         */
+        // Arrange
+        $range = new Range(-100, 100);   
+        $number = $this->randomNumber(
+            min: NextFloat::after($range->start->toFloat()),
+            max: NextFloat::before($range->end->toFloat())
+        );    
+        
+        // Act & Assert
+        $this->assertTrue(
+            $number->inRangeExtremesExcluded(new Range($range->start, $range->end)),
+            "$number is not inside $range->start/$range->end range."    
+        );
+
+        /**
+         * Out of range
+         */
+        // Arrange
+        $number = $this->randomNumber(max: $range->start->toFloat());
+
+        // Act & Assert
+        $this->assertFalse(
+            $number->inRangeExtremesExcluded($range),
+            "$number is inside $range->start/$range->end range."
+        );
+    }
+
     #[TestDox("can be converted to radian.")]
     public function test_toRadian(): void
     {
@@ -504,4 +638,5 @@ class NumberTest extends TestCaseWithDataProviders
     //     // Assert
     //     $this->assertEquals($log, $LOG, "log($arg, $base) = $LOG");
     // }
+
 }